// web/portal/client/src/entities/Schedule/ScheduleProperties.tsx
import { PropertySpec } from '@irontec-voip/ivoz-ui/services/api/ParsedApiSpecInterface';
import {
  EntityValue,
  EntityValues,
} from '@irontec-voip/ivoz-ui/services/entity/EntityService';

export type SchedulePropertyList<T> = {
  name?: T;
  timeIn?: T;
  timeout?: T;
  monday?: T;
  tuesday?: T;
  wednesday?: T;
  thursday?: T;
  friday?: T;
  saturday?: T;
  sunday?: T;
};

export type ScheduleProperties = SchedulePropertyList<Partial<PropertySpec>>;
export type SchedulePropertiesList = Array<
  SchedulePropertyList<EntityValue | EntityValues>
>;